Implement user types
User-side type registration, also known as a whole ass object system, implemented.  The manual contains the relevant updates with examples, but >TYPE and METH are the major additions.  As part of that, improved object copying for directories and builtins, in particular, and there's a new trivia constant CPDEPTH which keeps a typedir.dup() operation out of the cornfield.